Changes to support the release of GAMMA 4.2.1 which is the first followup release to our original GAMMA release six months ago.
- updated VERSION to 4.2.1

- In make/MakeMods --
-- Improved the definition of GAMMAVERSION. The old version was being 
re-evaluated as $(strip `cat ../../VERSION`) every time it was referenced 
which meant that it was difficult to use in subexpressions. (Not to mention
that the relative path was fragile.) Now it works the way I thought it worked
the first time I wrote it which is that it's evaluted just once.
-- Added code to parse the major, minor & release versions from GAMMAVERSION.
These replace a ton of hardcoded variables in the Makefiles.

- In platforms/XXXXX/Makefile, where XXXXX is *nix (OS X, Linux and Solaris) --
-- Deleted the unused and misleading PYVERSION
-- Replaced the hardcoded VERNO, MINNO and RELNO with MAJORVERSION,
MINORVERSION, and RELEASEVERSION from MakeMods. Now we don't have to touch 
every Makefile when the version changes. 

- Additionally, in platforms/OSX/Makefile --
-- Added a comment about building against 32-bit Python 2.5 on Snow Leopard
(see http://scion.duhs.duke.edu/vespa/gamma/ticket/13 )

- In platforms/XXXXX/gamma and gamma.in, where XXXXX is *nix (OS X, Linux 
and Solaris) --
-- Changed all the hardcoded version numbers from 4.2.0 to 4.2.1


After all these changes, I rebuilt PyGAMMA for all the platforms for which
we have prebuilt libs.



git-svn-id: http://scion.duhs.duke.edu/guest_svn/gamma/trunk@254 75907e46-c664-0410-ae3d-e599167a7a88