Adding pdf documentation, most of it originally in the gamma_docs repositoy. Kept same directory structure as what was there.
In two cases I reused the documentation that 
was originally located in the doc/pdf directory,
since only the subsection pdf's were availble
in the gamma_docs directory.

All the other "module" pdf's were confirmed to 
be the exact same size.

Effectively add sections on liquid nmr, strafi,
and the user manual.





git-svn-id: http://scion.duhs.duke.edu/guest_svn/gamma/trunk@288 75907e46-c664-0410-ae3d-e599167a7a88