A few improvements to the OS X build process.
First, on Matthias' suggestion, I changed the Makefile and .cc files to use
Apple-specific terms for referring to its BLAS/LAPACK libraries which allows
us to eliminate a long, version-specific path from the LAPACKINCL_FLAG.

Second, I added the variable MIN_OSX_VERSION to make it easy to build with
support for 10.5 (on my machine) and 10.6 (for most other people) as 
explained in the related comment in the Makefile.

Third, I added 'force' to the GAMMA script install target (see r390).




git-svn-id: http://scion.duhs.duke.edu/guest_svn/gamma/trunk@391 75907e46-c664-0410-ae3d-e599167a7a88