After screwing up the 4.3.2 bundle by putting Windows+Python 2.5 binaries in the Windows+Python 2.6 directory, I got frustrated and decided to automate an error-prone step of building the PyGAMMA download bundle. I had trouble keeping track of which binaries should be copied to which prebuilt libraries (and when). That's the step I automated.
In the OS X & Linux Makefiles I added a lot to the pysgdist target. In
addition to what it was already doing (copying build output to 
pygamma/homemade), it now also figures out the appropriate prebuilt directory 
(e.g. osx-32-py27) and copies the build output there, too. And, why not, it
automatically writes OS, compiler and Python version info the the 
build_notes.txt file there.

A lot of that code was cribbed from other places.

I didn't change the pysgdist in the Solaris Makefiles since we don't make
prebuilt libraries for Solaris.

In all the Makefiles, I changed the "Quick Info" comments at the top from a
numbered list to a bulleted list, since the numbers were a pain in the neck to
update and weren't really relevant.

In the OS X Makefile, I changed the code that detects whether or not Python
is 32- or 64-bit to a slightly more robust version.

Since Windows runs on a completely different build system, it got a 
different set of changes that do the same thing. The MSVC post-build step
used to run two shell commands (copy) and two Python scripts. I needed to
add a 3rd script but decided it would be easier just to merge them all into
one. The new script post_build.py subsumes clean_python_lib.py and 
write_pygamma_version.py and adds on the new feature of copying build output
to both homemade and the appropriate prebuilt lib. 

The Windows build is equivalent to the OS X and Linux build in this respect,
except that the Windows build doesn't write a build_notes.txt file just
because I can't be bothered at the moment to write the Python code to do it.



git-svn-id: http://scion.duhs.duke.edu/guest_svn/gamma/trunk@395 75907e46-c664-0410-ae3d-e599167a7a88