Fixed a small bug with the OS X & Linux builds. When building PyGAMMA (`make pysgdist`) to a prebuilt directory that didn't exist, the build would fail in a strange way. It would copy the binary to the directory name (e.g. linux-64-py27) and then try to copy files to that "directory".
The fix was to add an explicit check for the directory and create it if it
doesn't exist.

Didn't apply this change to the Solaris Makefiles since there's no section
for building the PyGAMMA distribution files in there.

The Windows build process is of course completely different but it made the
same flawed assumption so I fixed that too.


git-svn-id: http://scion.duhs.duke.edu/guest_svn/gamma/trunk@396 75907e46-c664-0410-ae3d-e599167a7a88